feat: 流水线组及流水线列表页优化 #7101

+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/service/view/PipelineViewGroupService.kt
@@ -150,7 +150,9 @@ class PipelineViewGroupService @Autowired constructor(
             val context = DSL.using(t)
             result = pipelineViewService.updateView(userId, projectId, viewId, pipelineView, context)
             if (result) {
-                pipelineViewGroupDao.remove(context, projectId, viewId)
+                if (pipelineView.pipelineIds != null) {
+                    pipelineViewGroupDao.remove(context, projectId, viewId)
+                }
                 redisOperation.delete(firstInitMark(projectId, viewId))
                 initViewGroup(
                     context = context,
@@ -323,7 +325,7 @@ class PipelineViewGroupService @Autowired constructor(
             watcher.stop()
         } else {
             watcher.start("initStaticViewGroup")
-            pipelineView.pipelineIds.forEach {
+            pipelineView.pipelineIds?.forEach {
                 pipelineViewGroupDao.create(
                     dslContext = context,
                     projectId = projectId,
@@ -424,7 +426,7 @@ class PipelineViewGroupService @Autowired constructor(
                 .filter { pipelineViewService.matchView(previewCondition, it) }
                 .map { it.pipelineId }
         } else {
-            pipelineView.pipelineIds.filter { allPipelineInfoMap.containsKey(it) }
+            pipelineView.pipelineIds?.filter { allPipelineInfoMap.containsKey(it) } ?: emptyList()
         }
 
         // 新增流水线 = 新流水线 - 老流水线
